Move C startup files into a separate folder
