+++ src/runtime/startup/c/wrapper.c
@@ -42,7 +42,7 @@
 #include "c/lib/util/message.h"
 #include "c/lib/util/util.h"
 
-#include "startup/wrapper.h"
+#include "startup/c/wrapper.h"
 
 // generated by GO libstartup build
 #include "lib/libstartup.h"
