+++ src/runtime/startup/scontainer.go
@@ -7,7 +7,7 @@ package main
 
 /*
 #include <sys/types.h>
-#include "startup/wrapper.h"
+#include "startup/c/wrapper.h"
 */
 // #cgo CFLAGS: -I../.. -I../../c/lib
 // #cgo LDFLAGS: -L../../../../builddir/lib/ -lruntime -luuid
